pyproject.toml
[tool.poetry]
name = "stream-man"
version = "0.1.0"
description = "Self hosted media tracking application."
authors = ["ryn-cx <84663588+ryn-cx@users.noreply.github.com>"]
license = "GNU AGPLv3"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
great-django-family = { git = "https://github.com/ryn-cx/great-django-family" }
html-file = { git = "https://github.com/ryn-cx/html-file" }
json-file = { git = "https://github.com/ryn-cx/json-file" }
pillow = "^10.1.0"
cryptography = "^41.0.7"
sorl-thumbnail = "^12.10.0"
playwright = "^1.40.0"
playwright-stealth = "^1.0.6"
django = "^5.0"
paved-path = { git = "https://github.com/ryn-cx/paved-path.git" }
keyring = "^25.0.0"


[tool.poetry.group.dev.dependencies]
ruff = "^0.1.9"
pre-commit = "^3.6.0"
pylint = "^3.0.3"
pytest = "^7.4.3"
django-types = "^0.19.1"
djlint = "^1.34.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.ruff]
line-length = 120

[tool.ruff.lint]
select = ["ALL"]
ignore = ["ANN101", "ANN102", "N804", "TD002", "TD003", "FIX002"]
# ANN101 - missing-type-self - self doesn't need type hints because the type is known implicitly
# ANN102 - missing-type-cls - cls doesn't need type hints because the type is known implicitly
# N804 - non-pep604-annotation - Sometimes has false positives, and Pylance is more accurate
# "TD002", "TD003", "FIX002 - Let me use TODO comments

[tool.ruff.extend-per-file-ignores]
"test_*.py" = ["S101", "INP001"]
# S101 - assert - Assert statements are fine in tests
# INP001 - implicit-namespace-package - Tests are not packages and should not have __init__.py files

[tool.ruff.lint.pydocstyle]
convention = "google"
